+++ actix-http/src/h1/dispatcher_tests.rs
@@ -15,7 +15,7 @@ use bytes::BytesMut;
 use futures_util::future::lazy;
 use tokio::time::{sleep, timeout};
 
-use super::dispatcher::{Dispatcher, DispatcherState, DispatcherStateProj, Flags};
+use super::*;
 use crate::{
     body::BoxBody,
     config::ServiceConfig,
@@ -1092,7 +1092,7 @@ async fn handler_drop_payload() {
         POST /drop-payload HTTP/1.1
         Host: localhost
         Content-Length: 3
-        
+
         abc
         ",
     ));
@@ -1155,7 +1155,7 @@ async fn handler_drop_payload() {
             POST /drop-payload HTTP/1.1
             Host: localhost
             Content-Length: 200
-            
+
             abc
             ",
         ));
@@ -1239,7 +1239,7 @@ async fn handler_drop_payload_drains_body() {
         POST /drop-payload HTTP/1.1
         Host: localhost
         Transfer-Encoding: chunked
-        
+
         ",
     ));
 
@@ -1529,7 +1529,7 @@ fn http_msg_creates_msg() {
             r"
             POST / HTTP/1.1
             Content-Length: 3
-            
+
             abc
             "
         ),
@@ -1541,7 +1541,7 @@ fn http_msg_creates_msg() {
             r"
             GET / HTTP/1.1
             Content-Length: 3
-            
+
             "
         ),
         "GET / HTTP/1.1\r\nContent-Length: 3\r\n\r\n"