+++ actix-http/src/h1/mod.rs
@@ -7,8 +7,6 @@ mod client;
 mod codec;
 mod decoder;
 mod dispatcher;
-#[cfg(test)]
-mod dispatcher_tests;
 mod encoder;
 mod expect;
 mod payload;